requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
bitcoin==1.1.42
celery==5.0.5
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.4
django-celery-beat==2.2.0
django-celery-results==2.1.0
django-email-verification==0.1.0
django-timezone-field==4.1.2
dnspython==1.16.0
eventlet==0.31.0
greenlet==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
kombu==5.1.0
Pillow==8.3.1
prompt-toolkit==3.0.19
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
qrcode==7.2
redis==3.5.3
requests==2.25.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
